[refactor] 더 일반적으로 사용되는 용어 활용하기
