+++ src/main/java/racingcar/config/MovingStrategyConfig.java
@@ -1,13 +1,13 @@
-package racingcar.setting;
+package racingcar.config;
 
-public enum MovingStrategySetting {
+public enum MovingStrategyConfig {
     MOVABLE_NUMBER(4),
     RANGE_START_NUMBER(0),
     RANGE_END_NUMBER(9);
 
     private final int setting;
 
-    MovingStrategySetting(final int setting) {
+    MovingStrategyConfig(final int setting) {
         this.setting = setting;
     }
