+++ src/main/java/racingcar/config/RaceRoundConfig.java
@@ -1,13 +1,13 @@
-package racingcar.setting;
+package racingcar.config;
 
-public enum RaceRoundSetting {
+public enum RaceRoundConfig {
 
     MIN_TOTAL_GAME_ROUND_COUNT(1),
     INIT_CURRENT_GAME_ROUND_COUNT(0);
 
     private final int setting;
 
-    RaceRoundSetting(final int setting) {
+    RaceRoundConfig(final int setting) {
         this.setting = setting;
     }
